requirements.txt
Django==4.2.29
asgiref==3.11.1
sqlparse==0.5.5
python-dotenv==0.21.0
Pillow==11.3.0
gunicorn==23.0.0
whitenoise==6.11.0
dj-database-url==3.0.1
psycopg2-binary==2.9.11
